+++ Witcher3MapViewer.Core/Interfaces/IMap.cs
