+++ Witcher3MapViewer.Core/Interfaces/IMapSettingsProvider.cs
