+++ Witcher3MapViewer.Core/Interfaces/IMarkerProvider.cs
